api/requirements.txt
boto3==1.26.137
pyjwt==2.6.0
stripe==5.4.0
redis==4.5.5
pydantic==1.10.7
email-validator==2.0.0
python-multipart==0.0.6
bcrypt==4.0.1
# uuid6==2023.5.2  # Not needed, using uuid4 instead
python-dateutil==2.8.2